+++ examples/data-sources/aiven_redis_user/data-source.tf
@@ -1,5 +1,0 @@
-data "aiven_redis_user" "user" {
-  service_name = "my-service"
-  project      = "my-project"
-  username     = "user1"
-}