+++ examples/resources/aiven_redis/import.sh
@@ -1,1 +1,0 @@
-terraform import aiven_redis.redis1 project/service_name