+++ internal/sdkprovider/service/redis/sweep.go
@@ -1,9 +1,0 @@
-package redis
-
-import (
-	"github.com/aiven/terraform-provider-aiven/internal/sweep"
-)
-
-func init() {
-	sweep.AddServiceSweeper("redis")
-}