+++ internal/sdkprovider/service/servicecomponent/service_component_data_source.go
@@ -41,7 +41,6 @@ func DatasourceServiceComponent() *schema.Resource {
 					"kafka_connect",
 					"mysql",
 					"pg",
-					"redis",
 					"kibana",
 					"kafka_connect",
 					"kafka_rest",