+++ internal/sdkprovider/service/vpc/sweep.go
@@ -19,7 +19,6 @@ func init() {
 		"aiven_pg",
 		"aiven_opensearch",
 		"aiven_grafana",
-		"aiven_redis",
 		"aiven_mysql",
 		"aiven_kafka",
 		"aiven_kafka_connect",